Apply changes from @DaNizz97 in EntityExcelSheet
