+++ core/src/main/java/com/creditdatamw/zerocell/handler/EntityExcelSheetHandler.java
@@ -5,60 +5,55 @@ import com.creditdatamw.zerocell.ZeroCellException;
 import com.creditdatamw.zerocell.ZeroCellReader;
 import com.creditdatamw.zerocell.column.ColumnInfo;
 import com.creditdatamw.zerocell.converter.Converter;
+import com.creditdatamw.zerocell.converter.Converters;
 import com.creditdatamw.zerocell.converter.NoopConverter;
-import org.apache.poi.hssf.util.CellReference;
 import org.apache.poi.ss.util.CellAddress;
+import org.apache.poi.ss.util.CellReference;
 import org.apache.poi.xssf.usermodel.XSSFComment;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.File;
 import java.lang.reflect.Field;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.Objects;
-
-import static com.creditdatamw.zerocell.converter.ConverterUtils.convertValueToType;
+import java.sql.Timestamp;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.util.*;
 
 final class EntityExcelSheetHandler<T> implements ZeroCellReader {
     private EntityHandler entityHandler;
     private final Logger LOGGER = LoggerFactory.getLogger(EntityExcelSheetHandler.class);
 
     private final ColumnInfo rowNumberColumn;
-    private final ColumnInfo[] columns;
+    private final Map<Integer, ColumnInfo> columns;
     private final List<T> entities;
     private final Converter NOOP_CONVERTER = new NoopConverter();
-    private final Converter[] converters;
-    private final int MAXIMUM_COL_INDEX;
-
+    private final Map<Integer, Converter> converters;
     private boolean isHeaderRow = false;
     private int currentRow = -1;
     private int currentCol = -1;
     private T cur;
 
-    EntityExcelSheetHandler(EntityHandler entityHandler, ColumnInfo rowNumberColumn, ColumnInfo[] columns) {
-        this.entityHandler = entityHandler;
+    EntityExcelSheetHandler(ColumnInfo rowNumberColumn, Map<Integer, ColumnInfo> columns) {
         this.rowNumberColumn = rowNumberColumn;
         this.columns = columns;
         this.converters = cacheConverters();
         this.entities = new ArrayList<>();
-        this.MAXIMUM_COL_INDEX = columns.length - 1;
     }
 
-    private Converter[] cacheConverters() {
-        Converter[] cv = new Converter[columns.length];
-        for (ColumnInfo c: columns) {
-            cv[c.getIndex()] = NOOP_CONVERTER;
+    private Map<Integer, Converter> cacheConverters() {
+        Map<Integer, Converter> mappedConverters = new HashMap<>();
+        for (ColumnInfo columnInfo : columns.values()) {
+            mappedConverters.put(columnInfo.getIndex(), NOOP_CONVERTER);
             try {
-                if (!c.getConverterClass().equals(NoopConverter.class)) {
-                    cv[c.getIndex()] = (Converter) c.getConverterClass().newInstance();
+                if (!columnInfo.getConverterClass().equals(NoopConverter.class)) {
+                    mappedConverters.put(columnInfo.getIndex(), (Converter) columnInfo.getConverterClass().newInstance());
                 }
             } catch (InstantiationException | IllegalAccessException e) {
-                LOGGER.error("Failed to instantiate Converter class: {}", c.getConverterClass());
+                LOGGER.error("Failed to instantiate Converter class: {}", columnInfo.getConverterClass());
             }
         }
-        return cv;
+        return mappedConverters;
     }
 
     @Override
@@ -93,7 +88,7 @@ final class EntityExcelSheetHandler<T> implements ZeroCellReader {
         try {
             cur = (T) entityHandler.getEntityClass().newInstance();
             // Write to the field with the @RowNumber annotation here if it exists
-            if (! Objects.isNull(rowNumberColumn)) {
+            if (!Objects.isNull(rowNumberColumn)) {
                 writeColumnField(cur, String.valueOf(currentRow), rowNumberColumn, currentRow);
             }
         } catch (InstantiationException | IllegalAccessException e) {
@@ -101,9 +96,18 @@ final class EntityExcelSheetHandler<T> implements ZeroCellReader {
         }
     }
 
+    private boolean isRowNumberValueSetted() {
+        return entityHandler.getMaxRowNumber() != 0 &&
+            entityHandler.getMaxRowNumber() != entityHandler.getSkipFirstNRows();
+    }
+
     @Override
     public void endRow(int i) {
-        if (! Objects.isNull(cur)) {
+        if (isRowNumberValueSetted() && i > entityHandler.getMaxRowNumber()) {
+            return;
+        }
+
+        if (!Objects.isNull(cur)) {
             this.entities.add(cur);
             cur = null;
         }
@@ -111,25 +115,19 @@ final class EntityExcelSheetHandler<T> implements ZeroCellReader {
 
     @Override
     public void cell(String cellReference, String formattedValue, XSSFComment xssfComment) {
-        // gracefully handle missing CellRef here in a similar way as XSSFCell does
-        if(cellReference == null) {
+        if (cellReference == null) {
             cellReference = new CellAddress(currentRow, currentCol).formatAsString();
         }
-
         int column = new CellReference(cellReference).getCol();
         currentCol = column;
 
-        // We ignore additional cells here since we only care about the cells
-        // in the columns array i.e. the defined columns
-        if (column > MAXIMUM_COL_INDEX) {
-            LOGGER.trace("Invalid Column index found: " + column);
+        ColumnInfo currentColumnInfo = columns.get(column);
+        if (Objects.isNull(currentColumnInfo)) {
             return;
         }
 
-        ColumnInfo currentColumnInfo = columns[column];
-
         if (isHeaderRow && !entityHandler.isSkipHeaderRow()) {
-            if (! currentColumnInfo.getName().equalsIgnoreCase(formattedValue.trim())){
+            if (!currentColumnInfo.getName().equalsIgnoreCase(formattedValue.trim())) {
                 throw new ZeroCellException(String.format("Expected Column '%s' but found '%s'", currentColumnInfo.getName(), formattedValue));
             }
         }
@@ -141,17 +139,17 @@ final class EntityExcelSheetHandler<T> implements ZeroCellReader {
     /**
      * Write the value read from the excel cell to a field
      *
-     * @param object the object to write to
-     * @param formattedValue the value read from the current excel column/row
+     * @param object            the object to write to
+     * @param formattedValue    the value read from the current excel column/row
      * @param currentColumnInfo Column metadata
-     * @param rowNum the row number
+     * @param rowNum            the row number
      */
     private void writeColumnField(T object, String formattedValue, ColumnInfo currentColumnInfo, int rowNum) {
         String fieldName = currentColumnInfo.getFieldName();
         try {
             Converter converter = NOOP_CONVERTER;
-            if (currentColumnInfo.getIndex() != -1 && currentColumnInfo.getIndex() < columns.length) {
-                converter = converters[currentColumnInfo.getIndex()];
+            if (currentColumnInfo.getIndex() != -1) {
+                converter = converters.get(currentColumnInfo.getIndex());
             }
             Object value = null;
             // Don't use a converter if there isn't a custom one
@@ -161,13 +159,14 @@ final class EntityExcelSheetHandler<T> implements ZeroCellReader {
                 // Handle any exceptions thrown by the converter - this stops execution of the whole process
                 try {
                     value = converter.convert(formattedValue, currentColumnInfo.getName(), rowNum);
-                } catch(Exception e) {
+                } catch (Exception e) {
                     throw new ZeroCellException(String.format("%s threw an exception while trying to convert value %s ", converter.getClass().getName(), formattedValue), e);
                 }
             }
-            Field field = entityHandler.getEntityClass().getDeclaredField(currentColumnInfo.getFieldName());
+            Field field = entityHandler.getEntityClass()
+                    .getDeclaredField(currentColumnInfo.getFieldName());
             boolean access = field.isAccessible();
-            if (! access) {
+            if (!access) {
                 field.setAccessible(true);
             }
             field.set(cur, value);
@@ -177,6 +176,40 @@ final class EntityExcelSheetHandler<T> implements ZeroCellReader {
         } catch (NoSuchFieldException | IllegalAccessException e) {
             LOGGER.error("Failed to set field: {}", fieldName, e);
         }
+    }
+
+    public Object convertValueToType(Class<?> fieldType, String formattedValue, String columnName, int rowNum) {
+        Object value = null;
+        if (fieldType == String.class) {
+            value = String.valueOf(formattedValue);
+        } else if (fieldType == LocalDateTime.class) {
+            return Converters.toLocalDateTime.convert(formattedValue, columnName, rowNum);
+
+        } else if (fieldType == LocalDate.class) {
+            return Converters.toLocalDate.convert(formattedValue, columnName, rowNum);
+
+        } else if (fieldType == java.sql.Date.class) {
+            return Converters.toSqlDate.convert(formattedValue, columnName, rowNum);
+
+        } else if (fieldType == Timestamp.class) {
+            return Converters.toSqlTimestamp.convert(formattedValue, columnName, rowNum);
+
+        } else if (fieldType == Integer.class || fieldType == int.class) {
+            return Converters.toInteger.convert(formattedValue, columnName, rowNum);
+
+        } else if (fieldType == Long.class || fieldType == long.class) {
+            return Converters.toLong.convert(formattedValue, columnName, rowNum);
+
+        } else if (fieldType == Double.class || fieldType == double.class) {
+            return Converters.toDouble.convert(formattedValue, columnName, rowNum);
+
+        } else if (fieldType == Float.class || fieldType == float.class) {
+            return Converters.toFloat.convert(formattedValue, columnName, rowNum);
+
+        } else if (fieldType == Boolean.class) {
+            return Converters.toBoolean.convert(formattedValue, columnName, rowNum);
+        }
+        return value;
     }
 
     @Override
